Bugfix FXIOS-4140 [v103] Remove old History Panel
