+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -983,7 +983,7 @@ class BrowserViewController: UIViewController {
 
         // Reset history panel pagination to get latest history visit
         if let historyPanel = libraryViewController.viewModel.panelDescriptors.first(where: {$0.panelType == .history}),
-           let vcPanel = historyPanel.viewController as? HistoryPanelWithGroups {
+           let vcPanel = historyPanel.viewController as? HistoryPanel {
             vcPanel.viewModel.shouldResetHistory = true
         }
