+++ Client/Frontend/Library/HistoryPanel/HistoryPanel.swift
@@ -1,18 +1,17 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/
-
-// TODO: HistoryPanel AND HistoryPanelV2 currently coexist until v100. Revert https://github.com/mozilla-mobile/firefox-ios/pull/10259 in v100.
-// Related to: https://mozilla-hub.atlassian.net/browse/FXIOS-2931 
+// file, You can obtain one at http://mozilla.org/MPL/2.0
 
 import UIKit
 import Shared
 import Storage
 import XCGLogger
 import WebKit
+import os.log
 
 private struct HistoryPanelUX {
     static let WelcomeScreenItemWidth = 170
+    static let HeaderHeight = CGFloat(40)
     static let IconSize = 23
     static let IconBorderColor = UIColor.Photon.Grey30
     static let IconBorderWidth: CGFloat = 0.5
@@ -26,226 +25,625 @@ private class FetchInProgressError: MaybeErrorType {
 }
 
 @objcMembers
-class HistoryPanel: SiteTableViewController, LibraryPanel {
+class HistoryPanel: UIViewController, LibraryPanel, Loggable, NotificationThemeable {
 
-    typealias HistoryPanelStrings = String.LibraryPanel.History
+    // MARK: - Properties
 
-    enum Section: Int {
-        // Showing recently closed, and clearing recent history are action rows of this type.
-        case additionalHistoryActions
-        case today
-        case yesterday
-        case lastWeek
-        case lastMonth
-        case older
-
-        static let count = 6
-
-        var title: String? {
-            switch self {
-            case .today:
-                return .LibraryPanel.Sections.Today
-            case .yesterday:
-                return .LibraryPanel.Sections.Yesterday
-            case .lastWeek:
-                return .LibraryPanel.Sections.LastWeek
-            case .lastMonth:
-                return .LibraryPanel.Sections.LastMonth
-            case .older:
-                return .LibraryPanel.Sections.Older
-            default:
-                return nil
-            }
-        }
-    }
-
-    enum AdditionalHistoryActionRow: Int {
-        case clearRecent
-        case showRecentlyClosedTabs
-
-        // Use to enable/disable the additional history action rows.
-        static func setStyle(enabled: Bool, forCell cell: OneLineTableViewCell) {
-            if enabled {
-                cell.titleLabel.alpha = 1.0
-                cell.leftImageView.alpha = 1.0
-                cell.selectionStyle = .default
-                cell.isUserInteractionEnabled = true
-            } else {
-                cell.titleLabel.alpha = 0.5
-                cell.leftImageView.alpha = 0.5
-                cell.selectionStyle = .none
-                cell.isUserInteractionEnabled = false
-            }
-        }
-    }
-
-    let QueryLimitPerFetch = 100
+    typealias HistoryPanelSections = HistoryPanelViewModel.Sections
+    typealias a11yIds = AccessibilityIdentifiers.LibraryPanels.HistoryPanel
 
     var libraryPanelDelegate: LibraryPanelDelegate?
     var recentlyClosedTabsDelegate: RecentlyClosedPanelDelegate?
 
-    var groupedSites = DateGroupedTableData<Site>()
-
-    var refreshControl: UIRefreshControl?
-
-    var currentFetchOffset = 0
-    var isFetchInProgress = false
-
-    var clearHistoryCell: OneLineTableViewCell?
+    let profile: Profile
+    let viewModel: HistoryPanelViewModel
     private let clearHistoryHelper: ClearHistoryHelper
+    var keyboardState: KeyboardState?
+    private lazy var siteImageHelper = SiteImageHelper(profile: profile)
+    var chevronImage = UIImage(named: ImageIdentifiers.menuChevron)
 
-    var hasRecentlyClosed: Bool {
-        return profile.recentlyClosedTabs.tabs.count > 0
+    // We'll be able to prefetch more often the higher this number is. But remember, it's expensive!
+    private let historyPanelPrefetchOffset = 8
+
+    var diffableDatasource: UITableViewDiffableDataSource<HistoryPanelSections, AnyHashable>?
+    private var hasRecentlyClosed: Bool { profile.recentlyClosedTabs.tabs.count > 0 }
+
+    // UI
+    var bottomStackView: BaseAlphaStackView = .build { _ in }
+
+    lazy var searchbar: UISearchBar = .build { searchbar in
+        searchbar.searchTextField.placeholder = self.viewModel.searchHistoryPlaceholder
+        searchbar.returnKeyType = .go
+        searchbar.delegate = self
     }
 
-    lazy var emptyStateOverlayView: UIView = createEmptyStateOverlayView()
+    lazy private var tableView: UITableView = .build { [weak self] tableView in
+        guard let self = self else { return }
+        tableView.dataSource = self.diffableDatasource
+        tableView.addGestureRecognizer(self.longPressRecognizer)
+        tableView.accessibilityIdentifier = a11yIds.tableView
+        tableView.prefetchDataSource = self
+        tableView.delegate = self
+        tableView.register(TwoLineImageOverlayCell.self, forCellReuseIdentifier: TwoLineImageOverlayCell.cellIdentifier)
+        tableView.register(TwoLineImageOverlayCell.self, forCellReuseIdentifier: TwoLineImageOverlayCell.accessoryUsageReuseIdentifier)
+        tableView.register(OneLineTableViewCell.self, forCellReuseIdentifier: OneLineTableViewCell.cellIdentifier)
+        tableView.register(SiteTableViewHeader.self, forHeaderFooterViewReuseIdentifier: SiteTableViewHeader.cellIdentifier)
+
+        if #available(iOS 15.0, *) {
+            tableView.sectionHeaderTopPadding = 0
+        }
+    }
 
     lazy var longPressRecognizer: UILongPressGestureRecognizer = {
-        return UILongPressGestureRecognizer(target: self, action: #selector(onLongPressGestureRecognized))
+        UILongPressGestureRecognizer(target: self, action: #selector(onLongPressGestureRecognized))
     }()
 
-    // MARK: - Lifecycle
+    lazy var emptyStateOverlayView: UIView = createEmptyStateOverlayView()
+    lazy var welcomeLabel: UILabel = .build { label in
+        label.text = self.viewModel.emptyStateText
+        label.textAlignment = .center
+        label.font = DynamicFontHelper.defaultHelper.DeviceFontLight
+        label.textColor = UIColor.theme.homePanel.welcomeScreenText
+        label.numberOfLines = 0
+        label.adjustsFontSizeToFitWidth = true
+    }
+    var refreshControl: UIRefreshControl?
+    var recentlyClosedCell: OneLineTableViewCell?
+
+    // MARK: - Inits
+
     init(profile: Profile, tabManager: TabManager) {
         self.clearHistoryHelper = ClearHistoryHelper(profile: profile, tabManager: tabManager)
-        super.init(profile: profile)
+        self.viewModel = HistoryPanelViewModel(profile: profile)
+        self.profile = profile
 
-        [ Notification.Name.FirefoxAccountChanged,
-          Notification.Name.PrivateDataClearedHistory,
-          Notification.Name.DynamicFontChanged,
-          Notification.Name.DatabaseWasReopened,
-          Notification.Name.OpenClearRecentHistory].forEach {
-            NotificationCenter.default.addObserver(self, selector: #selector(onNotificationReceived), name: $0, object: nil)
-        }
+        super.init(nibName: nil, bundle: nil)
     }
 
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 
+    deinit {
+        browserLog.debug("HistoryPanel Deinitialized.")
+    }
+
+    // MARK: - Lifecycle
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        tableView.addGestureRecognizer(longPressRecognizer)
-        tableView.accessibilityIdentifier = "History List"
-        tableView.prefetchDataSource = self
-        tableView.backgroundColor = UIColor.theme.homePanel.panelBackground
+
+        KeyboardHelper.defaultHelper.addDelegate(self)
+        viewModel.historyPanelNotifications.forEach {
+            NotificationCenter.default.addObserver(self, selector: #selector(handleNotifications), name: $0, object: nil)
+        }
+
+        handleRefreshControl()
+        setupLayout()
+        configureDatasource()
     }
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
-        // Add a refresh control if the user is logged in and the control was not added before. If the user is not
-        // logged in, remove any existing control.
-        if profile.hasSyncableAccount() && refreshControl == nil {
-            addRefreshControl()
-        } else if !profile.hasSyncableAccount() && refreshControl != nil {
-            removeRefreshControl()
+        bottomStackView.isHidden = !viewModel.isSearchInProgress
+        fetchDataAndUpdateLayout()
+    }
+
+    // MARK: - Private helpers
+
+    private func setupLayout() {
+        view.addSubview(tableView)
+        view.addSubview(bottomStackView)
+        bottomStackView.addArrangedSubview(searchbar)
+
+        NSLayoutConstraint.activate([
+            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            tableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+            tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
+
+            bottomStackView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            bottomStackView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+            bottomStackView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor)
+        ])
+    }
+
+    // Reload viewModel data and update layout
+    private func fetchDataAndUpdateLayout(animating: Bool = false) {
+        // Avoid refresing if search is in progress
+        guard !viewModel.isSearchInProgress else { return }
+
+        viewModel.reloadData { [weak self] success in
+            DispatchQueue.main.async {
+                self?.applySnapshot(animatingDifferences: animating)
+            }
         }
     }
 
-    // MARK: - Refreshing TableView
-    func addRefreshControl() {
-        let control = UIRefreshControl()
-        control.addTarget(self, action: #selector(onRefreshPulled), for: .valueChanged)
-        refreshControl = control
-        tableView.refreshControl = control
+    func updateLayoutForKeyboard() {
+        guard let keyboardHeight = keyboardState?.intersectionHeightForView(view),
+              keyboardHeight > 0 else {
+            bottomStackView.removeKeyboardSpacer()
+            return
+        }
+
+        let spacerHeight = keyboardHeight - UIConstants.BottomToolbarHeight
+        bottomStackView.addKeyboardSpacer(spacerHeight: spacerHeight)
+        bottomStackView.isHidden = false
     }
 
-    func removeRefreshControl() {
-        tableView.refreshControl = nil
-        refreshControl = nil
+    // Use to enable/disable the additional history action rows. `HistoryActionablesModel`
+    private func setTappableStateAndStyle(with item: AnyHashable, on cell: OneLineTableViewCell) {
+        var isEnabled = false
+
+        if let actionableItem = item as? HistoryActionablesModel {
+            switch actionableItem.itemIdentity {
+            case .clearHistory:
+                isEnabled = !viewModel.groupedSites.isEmpty
+            case .recentlyClosed:
+                isEnabled = hasRecentlyClosed
+                recentlyClosedCell = cell
+            default: break
+            }
+        }
+
+        // Set interaction behavior and style
+        cell.titleLabel.alpha = isEnabled ? 1.0 : 0.5
+        cell.leftImageView.alpha = isEnabled ? 1.0 : 0.5
+        cell.selectionStyle = isEnabled ? .default : .none
+        cell.isUserInteractionEnabled = isEnabled
     }
 
-    func endRefreshing() {
+    // MARK: - Datasource helpers
+
+    func siteAt(indexPath: IndexPath) -> Site? {
+        guard let siteItem = diffableDatasource?.itemIdentifier(for: indexPath) as? Site else { return nil }
+
+        return siteItem
+    }
+
+    private func showClearRecentHistory() {
+        clearHistoryHelper.showClearRecentHistory(onViewController: self, didComplete: { [weak self] date in
+            // Delete groupings that belong to THAT section.
+            if let date = date {
+                self?.viewModel.deleteGroupsForDates(date: date)
+            } else {
+                // Otherwise delete ALL groups, since we're deleting all history anyways.
+                self?.viewModel.removeAllData()
+            }
+
+            DispatchQueue.main.async {
+                self?.applySnapshot()
+                self?.tableView.reloadData()
+                self?.refreshRecentlyClosedCell()
+            }
+        })
+    }
+
+    private func refreshRecentlyClosedCell() {
+        guard let cell = recentlyClosedCell else { return }
+
+        self.setTappableStateAndStyle(
+            with: HistoryActionablesModel.activeActionables.first(where: { $0.itemIdentity == .recentlyClosed }),
+            on: cell)
+    }
+
+    func handleNotifications(_ notification: Notification) {
+        switch notification.name {
+        case .FirefoxAccountChanged, .PrivateDataClearedHistory:
+            viewModel.removeAllData()
+            fetchDataAndUpdateLayout(animating: true)
+
+            if profile.hasSyncableAccount() {
+                resyncHistory()
+            }
+            break
+        case .DynamicFontChanged:
+            if emptyStateOverlayView.superview != nil {
+                emptyStateOverlayView.removeFromSuperview()
+            }
+            emptyStateOverlayView = createEmptyStateOverlayView()
+            resyncHistory()
+            break
+        case .DatabaseWasReopened:
+            if let dbName = notification.object as? String, dbName == "browser.db" {
+                fetchDataAndUpdateLayout(animating: true)
+            }
+        case .OpenClearRecentHistory:
+            if viewModel.isSearchInProgress {
+                exitSearchState()
+            }
+
+            showClearRecentHistory()
+        default:
+            // no need to do anything at all
+            browserLog.error("Error: Received unexpected notification \(notification.name)")
+            break
+        }
+    }
+
+    // MARK: - UITableViewDataSource
+
+    /// Handles dequeuing the appropriate type of cell when needed.
+    private func configureDatasource() {
+        diffableDatasource = UITableViewDiffableDataSource<HistoryPanelSections, AnyHashable>(tableView: tableView) { [weak self] (tableView, indexPath, item) -> UITableViewCell? in
+            guard let self = self else {
+                Logger.browserLogger.error("History Panel - self became nil inside diffableDatasource!")
+                return nil
+            }
+
+            if let historyActionable = item as? HistoryActionablesModel {
+                guard let cell = tableView.dequeueReusableCell(withIdentifier: OneLineTableViewCell.cellIdentifier, for: indexPath) as? OneLineTableViewCell else {
+                    self.browserLog.error("History Panel - cannot create OneLineTableViewCell for historyActionable!")
+                    return nil
+                }
+
+                let actionableCell = self.configureHistoryActionableCell(historyActionable, cell)
+                return actionableCell
+            }
+
+            if let site = item as? Site {
+                guard let cell = tableView.dequeueReusableCell(withIdentifier: TwoLineImageOverlayCell.accessoryUsageReuseIdentifier, for: indexPath) as? TwoLineImageOverlayCell else {
+                    self.browserLog.error("History Panel - cannot create TwoLineImageOverlayCell for site!")
+                    return nil
+                }
+
+                let siteCell = self.configureSiteCell(site, cell)
+                return siteCell
+            }
+
+            if let searchTermGroup = item as? ASGroup<Site> {
+                guard let cell = tableView.dequeueReusableCell(withIdentifier: TwoLineImageOverlayCell.cellIdentifier, for: indexPath) as? TwoLineImageOverlayCell else {
+                    self.browserLog.error("History Panel - cannot create TwoLineImageOverlayCell for STG!")
+                    return nil
+                }
+
+                let asGroupCell = self.configureASGroupCell(searchTermGroup, cell)
+                return asGroupCell
+            }
+
+            // This should never happen! You will have an empty row!
+            return UITableViewCell()
+        }
+    }
+
+    private func configureHistoryActionableCell(_ historyActionable: HistoryActionablesModel, _ cell: OneLineTableViewCell) -> OneLineTableViewCell {
+        cell.titleLabel.text = historyActionable.itemTitle
+        cell.leftImageView.image = historyActionable.itemImage
+        cell.leftImageView.tintColor = .theme.browser.tint
+        cell.leftImageView.backgroundColor = .theme.homePanel.historyHeaderIconsBackground
+        cell.accessibilityIdentifier = historyActionable.itemA11yId
+        setTappableStateAndStyle(with: historyActionable, on: cell)
+
+        return cell
+    }
+
+    private func configureSiteCell(_ site: Site, _ cell: TwoLineImageOverlayCell) -> TwoLineImageOverlayCell {
+        cell.titleLabel.text = site.title
+        cell.titleLabel.isHidden = site.title.isEmpty
+        cell.descriptionLabel.text = site.url
+        cell.descriptionLabel.isHidden = false
+        cell.leftImageView.layer.borderColor = HistoryPanelUX.IconBorderColor.cgColor
+        cell.leftImageView.layer.borderWidth = HistoryPanelUX.IconBorderWidth
+        cell.accessoryView = nil
+        getFavIcon(for: site) { [weak cell] image in
+            cell?.leftImageView.image = image
+            cell?.leftImageView.backgroundColor = UIColor.theme.general.faviconBackground
+        }
+
+        return cell
+    }
+
+    private func getFavIcon(for site: Site, completion: @escaping (UIImage?) -> Void) {
+        siteImageHelper.fetchImageFor(site: site, imageType: .favicon, shouldFallback: false) { image in
+            completion(image)
+        }
+    }
+
+    private func configureASGroupCell(_ asGroup: ASGroup<Site>, _ cell: TwoLineImageOverlayCell) -> TwoLineImageOverlayCell {
+        if let groupCount = asGroup.description {
+            cell.descriptionLabel.text = groupCount
+        }
+
+        cell.titleLabel.text = asGroup.displayTitle
+        let imageView = UIImageView(image: chevronImage)
+        cell.accessoryView = imageView
+        cell.leftImageView.image = UIImage(named: ImageIdentifiers.stackedTabsIcon)?.withTintColor(ThemeManager.shared.currentTheme.colours.iconSecondary)
+        cell.leftImageView.backgroundColor = .theme.homePanel.historyHeaderIconsBackground
+
+        return cell
+    }
+
+    /// The data source gets populated here for your choice of section.
+    func applySnapshot(animatingDifferences: Bool = false) {
+        var snapshot = NSDiffableDataSourceSnapshot<HistoryPanelSections, AnyHashable>()
+
+        snapshot.appendSections(viewModel.visibleSections)
+
+        snapshot.sectionIdentifiers.forEach { section in
+            if !viewModel.hiddenSections.contains(where: { $0 == section }) {
+                snapshot.appendItems(viewModel.groupedSites.itemsForSection(section.rawValue - 1), toSection: section)
+            }
+        }
+
+        // Insert the ASGroup at the correct spot!
+        viewModel.searchTermGroups.forEach { grouping in
+            if let groupSection = viewModel.shouldAddGroupToSections(group: grouping) {
+                guard let individualItem = grouping.groupedItems.last, let lastVisit = individualItem.latestVisit else { return }
+
+                let groupTimeInterval = TimeInterval.fromMicrosecondTimestamp(lastVisit.date)
+
+                if let groupPlacedAfterItem = (viewModel.groupedSites.itemsForSection(groupSection.rawValue - 1)).first(where: { site in
+                    guard let lastVisit = site.latestVisit else { return false }
+                    return groupTimeInterval > TimeInterval.fromMicrosecondTimestamp(lastVisit.date)
+                }) {
+                    // In this case, we have Site items AND a group in the section.
+                    snapshot.insertItems([grouping], beforeItem: groupPlacedAfterItem)
+                } else {
+                    // Looks like this group's the only item in the section
+                    snapshot.appendItems([grouping], toSection: groupSection)
+                }
+            }
+        }
+
+        // Insert your fixed first section and data
+        if let historySection = snapshot.sectionIdentifiers.first, historySection != .additionalHistoryActions {
+            snapshot.insertSections([.additionalHistoryActions], beforeSection: historySection)
+        } else {
+            snapshot.appendSections([.additionalHistoryActions])
+        }
+        snapshot.appendItems(viewModel.historyActionables, toSection: .additionalHistoryActions)
+
+        diffableDatasource?.apply(snapshot, animatingDifferences: animatingDifferences, completion: nil)
+        updateEmptyPanelState()
+    }
+
+    // MARK: - Swipe Action helpers
+
+    func removeHistoryItem(at indexPath: IndexPath) {
+        guard let historyItem = diffableDatasource?.itemIdentifier(for: indexPath) else { return }
+
+        viewModel.removeHistoryItems(item: [historyItem], at: indexPath.section)
+
+        applySnapshot(animatingDifferences: true)
+    }
+
+    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+
+        // Adding support to delete item during search in next ticket
+        guard !viewModel.isSearchInProgress else { return nil }
+
+        // For UX consistency, every cell in history panel SHOULD have a trailing action.
+        let deleteAction = UIContextualAction(style: .destructive, title: .HistoryPanelDelete) { [weak self] (_, _, completion) in
+            guard let self = self else {
+                Logger.browserLogger.error("History Panel - self became nil inside SwipeActionConfiguration!")
+                completion(false)
+                return
+            }
+
+            self.removeHistoryItem(at: indexPath)
+        }
+
+        return UISwipeActionsConfiguration(actions: [deleteAction])
+    }
+
+    // MARK: - Empty State helpers
+
+    func updateEmptyPanelState() {
+        if viewModel.shouldShowEmptyState(searchText: searchbar.text ?? "") {
+            welcomeLabel.text = viewModel.emptyStateText
+            tableView.tableFooterView = emptyStateOverlayView
+        } else {
+            tableView.alwaysBounceVertical = true
+            tableView.tableFooterView = nil
+        }
+    }
+
+    private func createEmptyStateOverlayView() -> UIView {
+        let overlayView = UIView()
+
+        // overlayView becomes the footer view, and for unknown reason, setting the bgcolor is ignored.
+        // Create an explicit view for setting the color.
+        let bgColor: UIView = .build { view in
+            view.backgroundColor = UIColor.theme.homePanel.panelBackground
+        }
+        overlayView.addSubview(bgColor)
+
+        NSLayoutConstraint.activate([
+            bgColor.heightAnchor.constraint(equalToConstant: UIScreen.main.bounds.height),
+            bgColor.widthAnchor.constraint(equalTo: overlayView.widthAnchor)
+        ])
+
+        overlayView.addSubview(welcomeLabel)
+
+        let welcomeLabelPriority = UILayoutPriority(100)
+        NSLayoutConstraint.activate([
+            welcomeLabel.centerXAnchor.constraint(equalTo: overlayView.centerXAnchor),
+            welcomeLabel.centerYAnchor.constraint(equalTo: overlayView.centerYAnchor,
+                                                  constant: LibraryPanelUX.EmptyTabContentOffset).priority(welcomeLabelPriority),
+            welcomeLabel.topAnchor.constraint(greaterThanOrEqualTo: overlayView.topAnchor,
+                                              constant: 50),
+            welcomeLabel.widthAnchor.constraint(equalToConstant: CGFloat(HistoryPanelUX.WelcomeScreenItemWidth))
+        ])
+        return overlayView
+    }
+
+    // MARK: - Themeable
+
+    func applyTheme() {
+        updateEmptyPanelState()
+
+        tableView.backgroundColor = UIColor.theme.homePanel.panelBackground
+        searchbar.backgroundColor = UIColor.theme.textField.backgroundInOverlay
+        let tintColor = UIColor.theme.textField.textAndTint
+        let searchBarImage = UIImage(named: ImageIdentifiers.libraryPanelHistory)?.withRenderingMode(.alwaysTemplate).tinted(withColor: tintColor)
+        searchbar.setImage(searchBarImage, for: .search, state: .normal)
+        searchbar.tintColor = UIColor.theme.textField.textAndTint
+
+        tableView.reloadData()
+    }
+}
+
+// MARK: - UITableViewDelegate related helpers
+
+extension HistoryPanel: UITableViewDelegate {
+
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+
+        guard let item = diffableDatasource?.itemIdentifier(for: indexPath) else { return }
+
+        if let site = item as? Site {
+            handleSiteItemTapped(site: site)
+        }
+
+        if let historyActionable = item as? HistoryActionablesModel {
+            handleHistoryActionableTapped(historyActionable: historyActionable)
+        }
+
+        if let asGroupItem = item as? ASGroup<Site> {
+            handleASGroupItemTapped(asGroupItem: asGroupItem)
+        }
+    }
+
+    func scrollViewWillBeginDragging(_ scrollView: UIScrollView) {
+        if searchbar.isFirstResponder {
+            searchbar.resignFirstResponder()
+        }
+    }
+
+    private func handleSiteItemTapped(site: Site) {
+        guard let url = URL(string: site.url) else {
+            browserLog.error("Couldn't navigate to site: \(site.url)")
+            return
+        }
+
+        libraryPanelDelegate?.libraryPanel(didSelectURL: url, visitType: .typed)
+
+        TelemetryWrapper.recordEvent(category: .action,
+                                     method: .tap,
+                                     object: .selectedHistoryItem,
+                                     value: .historyPanelNonGroupItem,
+                                     extras: nil)
+    }
+
+    private func handleHistoryActionableTapped(historyActionable: HistoryActionablesModel) {
+        switch historyActionable.itemIdentity {
+        case .clearHistory:
+            showClearRecentHistory()
+        case .recentlyClosed:
+            navigateToRecentlyClosed()
+        default: break
+        }
+    }
+
+    private func handleASGroupItemTapped(asGroupItem: ASGroup<Site>) {
+        exitSearchState()
+
+        let asGroupListViewModel = SearchGroupedItemsViewModel(asGroup: asGroupItem, presenter: .historyPanel)
+        let asGroupListVC = SearchGroupedItemsViewController(viewModel: asGroupListViewModel, profile: profile)
+        asGroupListVC.libraryPanelDelegate = libraryPanelDelegate
+        asGroupListVC.title = asGroupItem.displayTitle
+
+        TelemetryWrapper.recordEvent(category: .action, method: .navigate, object: .navigateToGroupHistory, value: nil, extras: nil)
+
+        navigationController?.pushViewController(asGroupListVC, animated: true)
+    }
+
+    @objc private func sectionHeaderTapped(sender: UIGestureRecognizer) {
+        guard let sectionNumber = sender.view?.tag else {
+            return
+        }
+
+        viewModel.collapseSection(sectionIndex: sectionNumber)
+        applySnapshot()
+        // Needed to refresh the header state
+        tableView.reloadData()
+    }
+
+    // MARK: - TableView's Header & Footer view
+    func tableView(_ tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
+        if let header = view as? SiteTableViewHeader, let actualSection = viewModel.visibleSections[safe: section - 1] {
+
+            header.textLabel?.textColor = UIColor.theme.tableView.headerTextDark
+            header.contentView.backgroundColor = UIColor.theme.tableView.selectedBackground
+            header.textLabel?.text = actualSection.title // At worst, we have a header with no text.
+            header.collapsibleImageView.isHidden = false
+            let isCollapsed = viewModel.isSectionCollapsed(sectionIndex: section - 1)
+            header.collapsibleState = isCollapsed ? ExpandButtonState.right : ExpandButtonState.down
+
+            // Configure tap to collapse/expand section
+            header.tag = section
+            let tapGesture = UITapGestureRecognizer(target: self, action: #selector(sectionHeaderTapped(sender:)))
+            header.addGestureRecognizer(tapGesture)
+
+            // let historySectionsWithGroups
+            _ = viewModel.searchTermGroups.map { group in
+                viewModel.groupBelongsToSection(asGroup: group)
+            }
+        }
+
+    }
+
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        // First section is for recently closed and its header has no view.
+        guard HistoryPanelSections(rawValue: section) != .additionalHistoryActions else {
+            return nil
+        }
+
+        return tableView.dequeueReusableHeaderFooterView(withIdentifier: SiteTableViewHeader.cellIdentifier)
+    }
+
+    // viewForHeaderInSection REQUIRES implementing heightForHeaderInSection
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        // First section is for recently closed and its header has no height.
+        guard HistoryPanelSections(rawValue: section) != .additionalHistoryActions else {
+            return 0
+        }
+
+        return HistoryPanelUX.HeaderHeight
+    }
+}
+
+/// Refresh controls helpers
+extension HistoryPanel {
+    private func handleRefreshControl() {
+        if profile.hasSyncableAccount() && refreshControl == nil {
+            let control = UIRefreshControl()
+            control.addTarget(self, action: #selector(onRefreshPulled), for: .valueChanged)
+            refreshControl = control
+            tableView.refreshControl = control
+        } else if !profile.hasSyncableAccount() && refreshControl != nil {
+            tableView.refreshControl = nil
+            refreshControl = nil
+        }
+    }
+
+    private func endRefreshing() {
         // Always end refreshing, even if we failed!
         refreshControl?.endRefreshing()
 
         // Remove the refresh control if the user has logged out in the meantime
-        if !profile.hasSyncableAccount() {
-            removeRefreshControl()
-        }
+        handleRefreshControl()
     }
 
-    // MARK: - Loading data
-    override func reloadData() {
-        // Can be called while app backgrounded and the db closed, don't try to reload the data source in this case
-        guard !profile.isShutdown, !isFetchInProgress else { return }
-
-        fetchData().uponQueue(.main) { result in
-            if let sites = result.successValue {
-                let fetchedSites = sites.asArray()
-                let allCurrentGroupedSites = self.groupedSites.allItems()
-                let allUniquedSitesToAdd = (allCurrentGroupedSites + fetchedSites).uniqued().filter {
-                    !allCurrentGroupedSites.contains($0)
-                }
-
-                allUniquedSitesToAdd.forEach { site in
-                    if let latestVisit = site.latestVisit {
-                        self.groupedSites.add(site, timestamp: TimeInterval.fromMicrosecondTimestamp(latestVisit.date))
-                    }
-                }
-
-                self.tableView.reloadData()
-                self.updateEmptyPanelState()
-
-                if let cell = self.clearHistoryCell {
-                    AdditionalHistoryActionRow.setStyle(enabled: !self.groupedSites.isEmpty, forCell: cell)
-                }
-            }
-        }
-    }
-
-    func fetchData() -> Deferred<Maybe<Cursor<Site>>> {
-        guard !isFetchInProgress else {
-            return deferMaybe(FetchInProgressError())
-        }
-
-        isFetchInProgress = true
-
-        return profile.history.getSitesByLastVisit(limit: QueryLimitPerFetch, offset: currentFetchOffset) >>== { result in
-            // Force 100ms delay between resolution of the last batch of results
-            // and the next time `fetchData()` can be called.
-            DispatchQueue.main.asyncAfter(deadline: .now() + .milliseconds(100)) {
-                self.currentFetchOffset += self.QueryLimitPerFetch
-                self.isFetchInProgress = false
-            }
-
-            return deferMaybe(result)
-        }
-    }
-
-    func resyncHistory() {
+    private func resyncHistory() {
         profile.syncManager.syncHistory().uponQueue(.main) { syncResult in
             self.endRefreshing()
 
             if syncResult.isSuccess {
-                self.reloadData()
+                self.fetchDataAndUpdateLayout(animating: true)
             }
         }
     }
+}
 
-    // MARK: - Actions
-    func removeHistoryForURLAtIndexPath(indexPath: IndexPath) {
-        guard let site = siteForIndexPath(indexPath) else {
-            return
-        }
+/// User actions helpers
+extension HistoryPanel {
+    // MARK: - User Interactions
 
-        profile.history.removeHistoryForURL(site.url).uponQueue(.main) { result in
-            guard site == self.siteForIndexPath(indexPath) else {
-                self.reloadData()
-                return
-            }
-            self.tableView.beginUpdates()
-            self.groupedSites.remove(site)
-            self.tableView.deleteRows(at: [indexPath], with: .right)
-            self.tableView.endUpdates()
-            self.updateEmptyPanelState()
-
-            if let cell = self.clearHistoryCell {
-                AdditionalHistoryActionRow.setStyle(enabled: !self.groupedSites.isEmpty, forCell: cell)
-            }
-        }
-    }
-
+    /// When long pressed, a menu appears giving the choice of pinning as a Top Site.
     func pinToTopSites(_ site: Site) {
         profile.history.addPinnedTopSite(site).uponQueue(.main) { result in
             if result.isSuccess {
@@ -254,10 +652,8 @@ class HistoryPanel: SiteTableViewController, LibraryPanel {
         }
     }
 
-    func navigateToRecentlyClosed() {
-        guard hasRecentlyClosed else {
-            return
-        }
+    private func navigateToRecentlyClosed() {
+        guard hasRecentlyClosed else { return }
 
         let nextController = RecentlyClosedTabsPanel(profile: profile)
         nextController.title = .RecentlyClosedTabsPanelTitle
@@ -267,341 +663,34 @@ class HistoryPanel: SiteTableViewController, LibraryPanel {
         navigationController?.pushViewController(nextController, animated: true)
     }
 
-    private func showClearRecentHistory() {
-        clearHistoryHelper.showClearRecentHistory(onViewController: self, didComplete: { [weak self] _ in
-            self?.reloadData()
-        })
-    }
-
-    // MARK: - Cell configuration
-    func siteForIndexPath(_ indexPath: IndexPath) -> Site? {
-        // First section is reserved for recently closed.
-        guard indexPath.section > Section.additionalHistoryActions.rawValue else {
-            return nil
-        }
-
-        let sitesInSection = groupedSites.itemsForSection(indexPath.section - 1)
-        return sitesInSection[safe: indexPath.row]
-    }
-
-    func configureClearHistory(_ cell: OneLineTableViewCell, for indexPath: IndexPath) -> OneLineTableViewCell {
-        clearHistoryCell = cell
-        cell.titleLabel.text = HistoryPanelStrings.HistoryPanelClearHistoryButtonTitle
-        cell.leftImageView.image = UIImage.templateImageNamed("forget")
-        cell.leftImageView.tintColor = UIColor.theme.browser.tint
-        cell.leftImageView.backgroundColor = UIColor.theme.homePanel.historyHeaderIconsBackground
-        cell.accessibilityIdentifier = "HistoryPanel.clearHistory"
-
-        var isEmpty = true
-        for i in Section.today.rawValue..<tableView.numberOfSections {
-            if tableView.numberOfRows(inSection: i) > 0 {
-                isEmpty = false
-            }
-        }
-        AdditionalHistoryActionRow.setStyle(enabled: !isEmpty, forCell: cell)
-
-        return cell
-    }
-
-    func configureRecentlyClosed(_ cell: OneLineTableViewCell, for indexPath: IndexPath) -> OneLineTableViewCell {
-        cell.accessoryType = .disclosureIndicator
-        cell.titleLabel.text = HistoryPanelStrings.RecentlyClosedTabsButtonTitle
-        cell.leftImageView.image = UIImage.templateImageNamed("recently_closed")
-        cell.leftImageView.tintColor = UIColor.theme.browser.tint
-        cell.leftImageView.backgroundColor = UIColor.theme.homePanel.historyHeaderIconsBackground
-        AdditionalHistoryActionRow.setStyle(enabled: hasRecentlyClosed, forCell: cell)
-        cell.accessibilityIdentifier = "HistoryPanel.recentlyClosedCell"
-        return cell
-    }
-
-    func configureSite(_ cell: UITableViewCell, for indexPath: IndexPath) -> UITableViewCell {
-        if let site = siteForIndexPath(indexPath), let cell = cell as? TwoLineImageOverlayCell {
-            cell.titleLabel.text = site.title
-            cell.titleLabel.isHidden = site.title.isEmpty ? true : false
-            cell.descriptionLabel.isHidden = false
-            cell.descriptionLabel.text = site.url
-            cell.leftImageView.layer.borderColor = HistoryPanelUX.IconBorderColor.cgColor
-            cell.leftImageView.layer.borderWidth = HistoryPanelUX.IconBorderWidth
-            cell.leftImageView.contentMode = .center
-            cell.leftImageView.setImageAndBackground(forIcon: site.icon, website: site.tileURL) { [weak cell] in
-                cell?.leftImageView.image = cell?.leftImageView.image?.createScaled(CGSize(width: HistoryPanelUX.IconSize, height: HistoryPanelUX.IconSize))
-            }
-        }
-        return cell
-    }
-
-    // MARK: - Selector callbacks
-    func onNotificationReceived(_ notification: Notification) {
-        switch notification.name {
-        case .FirefoxAccountChanged, .PrivateDataClearedHistory:
-            groupedSites = DateGroupedTableData<Site>()
-            reloadData()
-
-            if profile.hasSyncableAccount() {
-                resyncHistory()
-            }
-            break
-        case .DynamicFontChanged:
-            reloadData()
-
-            if emptyStateOverlayView.superview != nil {
-                emptyStateOverlayView.removeFromSuperview()
-            }
-            emptyStateOverlayView = createEmptyStateOverlayView()
-            resyncHistory()
-            break
-        case .DatabaseWasReopened:
-            if let dbName = notification.object as? String, dbName == "browser.db" {
-                reloadData()
-            }
-        case .OpenClearRecentHistory:
-            showClearRecentHistory()
-        default:
-            // no need to do anything at all
-            print("Error: Received unexpected notification \(notification.name)")
-            break
-        }
-    }
-
-    func onLongPressGestureRecognized(_ longPressGestureRecognizer: UILongPressGestureRecognizer) {
+    @objc private func onLongPressGestureRecognized(_ longPressGestureRecognizer: UILongPressGestureRecognizer) {
         guard longPressGestureRecognizer.state == .began else { return }
         let touchPoint = longPressGestureRecognizer.location(in: tableView)
         guard let indexPath = tableView.indexPathForRow(at: touchPoint) else { return }
 
-        if indexPath.section != Section.additionalHistoryActions.rawValue {
+        if indexPath.section != HistoryPanelSections.additionalHistoryActions.rawValue {
             presentContextMenu(for: indexPath)
         }
     }
 
-    func onRefreshPulled() {
+    @objc private func onRefreshPulled() {
         refreshControl?.beginRefreshing()
         resyncHistory()
-    }
-
-    // MARK: - UITableViewDataSource
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return Section.count
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // First section is for recently closed and always has 1 row.
-        guard section > Section.additionalHistoryActions.rawValue else {
-            return 2
-        }
-
-        return groupedSites.numberOfItemsForSection(section - 1)
-    }
-
-    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        // First section is for recently closed and has no title.
-        guard section > Section.additionalHistoryActions.rawValue else {
-            return nil
-        }
-
-        // Ensure there are rows in this section.
-        guard groupedSites.numberOfItemsForSection(section - 1) > 0 else {
-            return nil
-        }
-
-        return Section(rawValue: section)?.title
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = super.tableView(tableView, cellForRowAt: indexPath) as! TwoLineImageOverlayCell
-        cell.accessoryType = .none
-        // First section is reserved for recently closed.
-        guard indexPath.section > Section.additionalHistoryActions.rawValue else {
-            cell.leftImageView.layer.borderWidth = 0
-
-            guard let row = AdditionalHistoryActionRow(rawValue: indexPath.row) else {
-                assertionFailure("Bad row number")
-                return cell
-            }
-
-            let oneLineCell = tableView.dequeueReusableCell(withIdentifier: OneLineCellIdentifier, for: indexPath) as! OneLineTableViewCell
-            switch row {
-            case .clearRecent:
-                return configureClearHistory(oneLineCell, for: indexPath)
-            case .showRecentlyClosedTabs:
-                return configureRecentlyClosed(oneLineCell, for: indexPath)
-            }
-        }
-
-        return configureSite(cell, for: indexPath)
-    }
-
-    // MARK: - UITableViewDelegate
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        // First section is reserved for recently closed.
-        defer {
-            tableView.deselectRow(at: indexPath, animated: true)
-        }
-        guard indexPath.section > Section.additionalHistoryActions.rawValue else {
-            switch indexPath.row {
-            case 0:
-                clearHistoryHelper.showClearRecentHistory(onViewController: self, didComplete: { [weak self] _ in
-                    self?.reloadData()
-                })
-                showClearRecentHistory()
-
-            default:
-                navigateToRecentlyClosed()
-            }
-            return
-        }
-
-        if let site = siteForIndexPath(indexPath), let url = URL(string: site.url) {
-            if let libraryPanelDelegate = libraryPanelDelegate {
-                libraryPanelDelegate.libraryPanel(didSelectURL: url, visitType: VisitType.typed)
-            }
-            return
-        }
-        print("Error: No site or no URL when selecting row.")
-    }
-
-    override func tableView(_ tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
-        if let header = view as? UITableViewHeaderFooterView {
-            header.textLabel?.textColor = UIColor.theme.tableView.headerTextDark
-            header.contentView.backgroundColor = UIColor.theme.tableView.selectedBackground
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        // First section is for recently closed and its header has no view.
-        guard section > Section.additionalHistoryActions.rawValue else {
-            return nil
-        }
-
-        // Ensure there are rows in this section.
-        guard groupedSites.numberOfItemsForSection(section - 1) > 0 else {
-            return nil
-        }
-
-        return super.tableView(tableView, viewForHeaderInSection: section)
-    }
-
-    override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        // First section is for recently closed and its header has no height.
-        guard section > Section.additionalHistoryActions.rawValue else {
-            return 0
-        }
-
-        // Ensure there are rows in this section.
-        guard groupedSites.numberOfItemsForSection(section - 1) > 0 else {
-            return 0
-        }
-
-        return super.tableView(tableView, heightForHeaderInSection: section)
-    }
-
-    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        guard indexPath.section != Section.additionalHistoryActions.rawValue else {
-            return nil
-        }
-
-        let deleteAction = UIContextualAction(style: .destructive, title: .HistoryPanelDelete) { [weak self] (_, _, completion) in
-            guard let strongSelf = self else { completion(false); return }
-
-            strongSelf.removeHistoryForURLAtIndexPath(indexPath: indexPath)
-            completion(true)
-        }
-
-        return UISwipeActionsConfiguration(actions: [deleteAction])
-    }
-
-    // MARK: - Empty State
-    func updateEmptyPanelState() {
-        if groupedSites.isEmpty {
-            if emptyStateOverlayView.superview == nil {
-                tableView.tableFooterView = emptyStateOverlayView
-            }
-        } else {
-            tableView.alwaysBounceVertical = true
-            tableView.tableFooterView = nil
-        }
-    }
-
-    func createEmptyStateOverlayView() -> UIView {
-        let overlayView = UIView()
-
-        // overlayView becomes the footer view, and for unknown reason, setting the bgcolor is ignored.
-        // Create an explicit view for setting the color.
-        let bgColor = UIView()
-        bgColor.backgroundColor = UIColor.theme.homePanel.panelBackground
-        overlayView.addSubview(bgColor)
-        bgColor.snp.makeConstraints { make in
-            // Height behaves oddly: equalToSuperview fails in this case, as does setting top.equalToSuperview(), simply setting this to ample height works.
-            make.height.equalTo(UIScreen.main.bounds.height)
-            make.width.equalToSuperview()
-        }
-
-        let welcomeLabel = UILabel()
-        overlayView.addSubview(welcomeLabel)
-        welcomeLabel.text = .HistoryPanelEmptyStateTitle
-        welcomeLabel.textAlignment = .center
-        welcomeLabel.font = DynamicFontHelper.defaultHelper.DeviceFontLight
-        welcomeLabel.textColor = UIColor.theme.homePanel.welcomeScreenText
-        welcomeLabel.numberOfLines = 0
-        welcomeLabel.adjustsFontSizeToFitWidth = true
-
-        welcomeLabel.snp.makeConstraints { make in
-            make.centerX.equalTo(overlayView)
-            // Sets proper top constraint for iPhone 6 in portait and for iPad.
-            make.centerY.equalTo(overlayView).offset(LibraryPanelUX.EmptyTabContentOffset).priority(100)
-            // Sets proper top constraint for iPhone 4, 5 in portrait.
-            make.top.greaterThanOrEqualTo(overlayView).offset(50)
-            make.width.equalTo(HistoryPanelUX.WelcomeScreenItemWidth)
-        }
-        return overlayView
-    }
-
-    override func applyTheme() {
-        emptyStateOverlayView.removeFromSuperview()
-        emptyStateOverlayView = createEmptyStateOverlayView()
-        updateEmptyPanelState()
-
-        super.applyTheme()
     }
 }
 
 extension HistoryPanel: UITableViewDataSourcePrefetching {
-    func tableView(_ tableView: UITableView, prefetchRowsAt indexPaths: [IndexPath]) {
-        guard !isFetchInProgress, indexPaths.contains(where: shouldLoadRow) else { return }
 
-        reloadData()
+    // Happens WAY too often. We should consider fetching the next set when the user HITS the bottom instead.
+    func tableView(_ tableView: UITableView, prefetchRowsAt indexPaths: [IndexPath]) {
+        guard !viewModel.isFetchInProgress, indexPaths.contains(where: shouldLoadRow) else { return }
+
+        fetchDataAndUpdateLayout(animating: false)
     }
 
     func shouldLoadRow(for indexPath: IndexPath) -> Bool {
-        guard indexPath.section > Section.additionalHistoryActions.rawValue else {
-            return false
-        }
+        guard HistoryPanelSections(rawValue: indexPath.section) != .additionalHistoryActions else { return false }
 
-        return indexPath.row >= groupedSites.numberOfItemsForSection(indexPath.section - 1) - 1
-    }
-}
-
-extension HistoryPanel: LibraryPanelContextMenu {
-    func presentContextMenu(for site: Site, with indexPath: IndexPath, completionHandler: @escaping () -> PhotonActionSheet?) {
-        guard let contextMenu = completionHandler() else { return }
-        present(contextMenu, animated: true, completion: nil)
-    }
-
-    func getSiteDetails(for indexPath: IndexPath) -> Site? {
-        return siteForIndexPath(indexPath)
-    }
-
-    func getContextMenuActions(for site: Site, with indexPath: IndexPath) -> [PhotonRowActions]? {
-        guard var actions = getDefaultContextMenuActions(for: site, libraryPanelDelegate: libraryPanelDelegate) else { return nil }
-
-        let removeAction = SingleActionViewModel(title: .DeleteFromHistoryContextMenuTitle, iconString: "action_delete", tapHandler: { _ in
-            self.removeHistoryForURLAtIndexPath(indexPath: indexPath)
-        })
-
-        let pinTopSite = SingleActionViewModel(title: .AddToShortcutsActionTitle, iconString: ImageIdentifiers.addShortcut, tapHandler: { _ in
-            self.pinToTopSites(site)
-        })
-        actions.append(PhotonRowActions(pinTopSite))
-        actions.append(PhotonRowActions(removeAction))
-        return actions
+        return indexPath.row >= viewModel.groupedSites.numberOfItemsForSection(indexPath.section - 1) - historyPanelPrefetchOffset
     }
 }
